requirements.txt
segmentation-models-pytorch
six
albumentations